Allow differentiating between terrain bashed from the side and above
Add ter_set_bashed_from_above to bash data, and default it to ter_set.

This allows terrain to define the terrain it should be transformed into
when it is bashed from above as opposed to from the side, as sometimes
it makes sense for that to be different (e.g. when bashing into t_rock,
which normally transforms into t_rock_floor, which has a roof of
t_flat_roof--this makes sense when bashing from the side, but from above
there shouldn't be a solid roof since you just bashed through it to get
to the t_rock you bashed).

Add new t_rock_floor_no_roof which has t_open_air for a roof, and use it
as the ter_set_bashed_from_above for t_rock.

The end result is when experimental z-levels are on and you smash the
floor of your current z-level through into the rock a z-level below you
and destroy that too, instead of a t_flat_roof magically appearing at
your z-level despite the terrain below being hollow, you instead get
t_open_air so you can climb/fall down into the hole you made.