Make bash_below false for some terrain types
Bash below is poorly documented (just "TODO" in the json documentation)
but the comment in the source code says that:

// This terrain is the roof of the tile below it, try to destroy that
too

I think this is inappropriately applied to some terrain types, such as
t_pavement, because it means that if you have experimental z-levels
enabled and you use a pickaxe to destroy a piece of road pavement,
instead of breaking up the pavement, exposing the underlying terrain and
leaving behind the debris byproducts, you instead dig an entire z-level
down below your current z-level and mine out the rock there too.

The pavement is not the roof of the rock below it, and in general where
this flag was set to true for something that is actually more of a layer
on top of the ground (e.g. pavement, woodchips, sidewalk, wood floor)
I've removed it.